[MISC] move include/seqan3/{core,utility}/simd/{,detail/}debug_stream_simd.hpp
