+++ include/seqan3/core/simd/all.hpp
@@ -32,7 +32,7 @@
  */
 
 #include <seqan3/utility/simd/concept.hpp>
-#include <seqan3/core/simd/debug_stream_simd.hpp>
+#include <seqan3/utility/simd/detail/debug_stream_simd.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
 #include <seqan3/utility/simd/simd.hpp>
