+++ test/unit/core/simd/CMakeLists.txt
@@ -1,3 +1,1 @@
 add_subdirectories()
-
-seqan3_test(debug_stream_simd_test.cpp)
