+++ test/unit/utility/simd/detail/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_test (debug_stream_simd_test.cpp)
